chore: add angular value bindings
