+++ src/components/atoms/modus-wc-tooltip/modus-wc-tooltip.scss
